feature(proctree): create process tree processors
1. Make the pipeline events to interact with process tree and vice-versa.
2. Introduce a Setuid() processor for the process tree (example for the rest).
